Create ToDolist.cpp
Interface is presented in Russian. 
Code operation algorithm:
1)A Todo structure and a todoList vector are created, representing the to-dos and the to-do list, respectively.

2)5 sample cases are initialized by setting their field values.

3)The main program loop begins where the user is prompted to select an action from a menu.

4)Depending on the user's choice, the appropriate function is called:
Adding a To-Do: prompts the user for the name, priority, description, and date and time of the to-do, which are then added to the to-do list.
Deleting a Case: The user is prompted to select a case from the list by its index, which is then removed from the list.
Editing a case: the user is prompted to select a case from the list by its index and then enter new field values for editing.
Searching by title, priority, description, and due date and time: The user is prompted to enter an appropriate query, and the function displays cases that match the search criteria.
Displaying sorted to-do list: the user is prompted to select sorting by priority or by due date and time, after which the to-do list is displayed.
Program exit: The loop is interrupted and the program ends.

5)After performing the selected action, the program returns to step 3 and waits for a new user selection.

This is the main algorithm of the code that provides the functionality of the "To Do List" with the ability to add, delete, edit, search and display cases.